Add server-side python capabilities.
In order to remotely execute fabric tasks on the server, we've added python's
virtualenv to the server, and we create a default virtualenv with some basic
requirements so Otto can do his thing without disturbing the system python.

Also, abstracted the hard-coded otto path that was being used in multiple
spots.