Use git better, run build at server.
This branch is starting to shift some of Otto to the server, using git
to push files up, and running the build task at the server. Now that the
source and the virtualenv will exist at the server, it should be much
easier to create automated processes that update the repository.